requirements.txt
django==1.8.2
Pillow==2.8.2
sorl-thumbnail==12.2
django-polymorphic==0.7.1
